+++ lib/active_graphql/client/actions/action/format_inputs.rb
@@ -22,7 +22,7 @@ module ActiveGraphql
 
           def formatted(attributes)
             if attributes.is_a?(Hash)
-              attributes.map { |key, val| formatted_key_and_value(key, val) }.join(', ')
+              formatted_attributes(attributes)
             else
               raise(
                 ActiveGraphql::Client::Actions::WrongTypeError,
@@ -31,8 +31,31 @@ module ActiveGraphql
             end
           end
 
+          def formatted_attributes(attributes)
+            attributes = attributes.dup
+            keyword_fields = (attributes.delete(:__keyword_attributes) || []).map(&:to_s)
+
+            formatted_attributes = attributes.map do |key, val|
+              if keyword_fields.include?(key.to_s)
+                formatted_key_and_keyword(key, val)
+              else
+                formatted_key_and_value(key, val)
+              end
+            end
+
+            formatted_attributes.join(', ')
+          end
+
           def formatted_key_and_value(key, value)
             "#{key}: #{formatted_value(value)}"
+          end
+
+          def formatted_key_and_keyword(key, value)
+            if value.is_a?(String) || value.is_a?(Symbol)
+              "#{key}: #{value}"
+            else
+              "#{key}: #{formatted_value(value)}"
+            end
           end
 
           def formatted_value(value) # rubocop:disable Metrics/MethodLength